test: tests for the logger belong to the common module
