+++ test/integration_tests/IO/CMakeLists.txt
@@ -36,7 +36,6 @@ list(APPEND TESTS "test_IO_00003")
 list(APPEND TESTS "test_IO_00004")
 list(APPEND TESTS "test_IO_00005")
 list(APPEND TESTS "test_IO_00006")
-list(APPEND TESTS "test_IO_00007")
 
 # Test extra libraries
 set(TEST_EXTRA_LIBRARIES "")
@@ -53,6 +52,11 @@ add_custom_command(
 )
 
 add_custom_command(
+    TARGET "integration_test_IO_00002" PRE_BUILD
+    COMMAND ${CMAKE_COMMAND} -E copy_if_different "${CMAKE_CURRENT_SOURCE_DIR}/data/configuration.json" "${CMAKE_CURRENT_BINARY_DIR}/data/configuration.json"
+)
+
+add_custom_command(
     TARGET "integration_test_IO_00004" PRE_BUILD
     COMMAND ${CMAKE_COMMAND} -E copy_if_different "${CMAKE_CURRENT_SOURCE_DIR}/data/configuration.xml" "${CMAKE_CURRENT_BINARY_DIR}/data/configuration.xml"
 )
@@ -62,9 +66,4 @@ add_custom_command(
     COMMAND ${CMAKE_COMMAND} -E copy_if_different "${CMAKE_CURRENT_SOURCE_DIR}/data/cubeAscii.stl" "${CMAKE_CURRENT_BINARY_DIR}/data/cubeAscii.stl"
     COMMAND ${CMAKE_COMMAND} -E copy_if_different "${CMAKE_CURRENT_SOURCE_DIR}/data/cubeBinary.stl" "${CMAKE_CURRENT_BINARY_DIR}/data/cubeBinary.stl"
     COMMAND ${CMAKE_COMMAND} -E copy_if_different "${CMAKE_CURRENT_SOURCE_DIR}/data/cubeAscii_MultiSolid.stl" "${CMAKE_CURRENT_BINARY_DIR}/data/cubeAscii_MultiSolid.stl"
-)
-
-add_custom_command(
-    TARGET "integration_test_IO_00007" PRE_BUILD
-    COMMAND ${CMAKE_COMMAND} -E copy_if_different "${CMAKE_CURRENT_SOURCE_DIR}/data/configuration.json" "${CMAKE_CURRENT_BINARY_DIR}/data/configuration.json"
 )
