+++ test/integration_tests/IO/test_IO_00002.cpp
@@ -22,12 +22,10 @@
  *
 \*---------------------------------------------------------------------------*/
 
-#include <array>
 #if BITPIT_ENABLE_MPI==1
 #include <mpi.h>
 #endif
 
-#include "bitpit_common.hpp"
 #include "bitpit_IO.hpp"
 
 using namespace bitpit;
@@ -35,164 +33,69 @@ using namespace bitpit;
 /*!
 * Subtest 001
 *
-* Testing basic logger fatures.
+* Testing json parser.
+* BEWARE: if bitpit has no JSON support (RapidJSON package not installed)
+* this test does nothing and return .
 */
 int subtest_001()
 {
-	int nProcessors;
-	int rank;
+    std::cout << "Testing json configuraton parser features" << std::endl;
 
-#if BITPIT_ENABLE_MPI==1
-	MPI_Comm_size(MPI_COMM_WORLD, &nProcessors);
-	MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-#else
-	nProcessors = 1;
-	rank        = 0;
-#endif
+    // Declare a bitpit config parser with multisections enabled.
+    //
+    config::reset("bitpit", 1, true);
 
-	std::cout << "Testing basic logger fatures" << "\n";
+    // Read the configuration file
+    std::cout << std::endl;
+    std::cout << "Read json configuration file..." << std::endl;
 
-	// Default logger
-	log::manager().initialize(log::SEPARATE, false, nProcessors, rank);
+    config::read("data/configuration.json");
 
-	log::cout() << consoleVerbosity(log::NORMAL);
-	log::cout() << fileVerbosity(log::NORMAL);
+    bitpit::GlobalConfigParser & root = config::root;
+    // Dump the configuration
+    std::cout << std::endl;
+    std::cout << "Dump configuration..." << std::endl;
+    root.dump(std::cout, 1);
 
-	log::cout() << log::priority(log::NORMAL);
-	log::cout() << "1-1\n2--2\n" << std::endl << "3---3" << std::endl;
-	log::cout() << "4----4\n5-----5\n" << std::endl << "6--";
-	log::cout() << "----6\n7-------7";
-	log::cout() << std::endl;
-	log::cout() << log::priority(log::NORMAL);
-	log::cout() << log::context("context-1");
-	log::cout() << "1-1\n2--2\n" << std::endl << "3---3" << std::endl;
-	log::cout() << "4----4\n5-----5\n" << std::endl << "6--";
-	log::cout() << "----6\n7-------7";
-	log::cout() << log::priority(log::NORMAL);
-	log::cout() << log::context("context-2");
-	log::cout() << "1-1\n2--2\n" << std::endl << "3---3" << std::endl;
-	log::cout() << "4----4\n5-----5\n" << std::endl << "6--";
-	log::cout() << "----6\n7-------7";
-	log::cout() << std::endl;
+    // Access configuration
+    std::cout << std::endl;
+    std::cout << "Access configuration..." << std::endl;
+    std::cout << "  - Section \"first\" has color..." << root["first"].get("color") << std::endl;
+    std::cout << "  - Section \"second\" has color..." << root["second"].get("color") << std::endl;
+    std::cout << "  - Section \"first\" has distance..." << root.getSection("first").get("distance") << std::endl;
+    std::cout << "  - Section \"second\" has y data..." << root["second"]["data"].get("y") << std::endl;
+    std::cout << "  - Section \"first\" option count..." << root.getSection("first").getOptionCount() << std::endl;
+    std::cout << "  - Section \"first\" sub-section count..." << root.getSection("first").getSectionCount() << std::endl;
 
-	BITPIT_DEBUG_COUT() << "Debug statement" << std::endl;
+    int firstDistanceInt = root["first"].get<int>("distance");
+    std::cout << "  - Section \"first\" has distance (int)..." << firstDistanceInt << std::endl;
 
-	// Log only on console
-	log::manager().create("logger1", false, nProcessors, rank);
+    double firstDistanceDouble = root["first"].get<double>("distance");
+    std::cout << "  - Section \"first\" has distance (double)..." << firstDistanceDouble << std::endl;
 
-	log::cout("logger1") << consoleVerbosity(log::DEBUG);
-	log::cout("logger1") << fileVerbosity(log::QUIET);
+    double secondDataDouble = root["second"]["data"].get<double>("y");
+    std::cout << "  - Section \"second\" has y data (double)..." << secondDataDouble << std::endl;
 
-	log::cout("logger1") << log::priority(log::DEBUG);
-	log::cout("logger1") << log::context("logger1-A");
-	log::cout("logger1") << "1-1\n2--2\n" << std::endl << "3---3" << std::endl;
-	log::cout("logger1") << "4----4\n5-----5\n" << std::endl << "6--";
-	log::cout("logger1") << "----6\n7-------7";
-	log::cout("logger1") << std::endl;
-	log::cout("logger1") << log::priority(log::NORMAL);
-	log::cout("logger1") << log::context("logger1-B");
-	log::cout("logger1") << "1-1\n2--2\n" << std::endl << "3---3" << std::endl;
-	log::cout("logger1") << "4----4\n5-----5\n" << std::endl << "6--";
-	log::cout("logger1") << "----6\n7-------7";
-	log::cout("logger1") << std::endl;
+    bool firstExistsBool = root["first"].get<bool>("exists");
+    std::cout << "  - Section \"first\" has exists..." << firstExistsBool << std::endl;
 
-	BITPIT_DEBUG_COUT("logger1") << "Debug statement" << std::endl;
+    std::cout << "  - Empty dummy option reading ... -" << root.getSection("first").get("dummy")<<" - "<< std::endl;
 
-	// Log on console and also on file
-	log::manager().create("logger2", false, nProcessors, rank);
+    std::cout << "  - Non existent option with fallback..." << root.getSection("first").get("none", "111") << std::endl;
+    std::cout << "  - Non existent option with fallback (int)..." << root.getSection("first").get<int>("none", 111) << std::endl;
+    std::cout << "  - Non existent option with fallback (double)..." << root.getSection("first").get<double>("none", 111.111) << std::endl;
 
-	log::cout("logger2") << consoleVerbosity(log::NORMAL);
-	log::cout("logger2") << fileVerbosity(log::NORMAL);
 
-	log::cout("logger2") << log::priority(log::DEBUG);
-	log::cout("logger2") << log::context("logger2-A");
-	log::cout("logger2") << "1-1\n2--2\n" << std::endl << "3---3" << std::endl;
-	log::cout("logger2") << "4----4\n5-----5\n" << std::endl << "6--";
-	log::cout("logger2") << "----6\n7-------7";
-	log::cout("logger2") << std::endl;
-	log::cout("logger2") << log::priority(log::NORMAL);
-	log::cout("logger2") << log::context("logger2-B");
-	log::cout("logger2") << "1-1\n2--2\n" << std::endl << "3---3" << std::endl;
-	log::cout("logger2") << "4----4\n5-----5\n" << std::endl << "6--";
-	log::cout("logger2") << "----6\n7-------7";
-	log::cout("logger2") << std::endl;
+    Config::MultiSection multisections = root.getSections("ObjArray");
+    for(auto sec : multisections){
+        std::cout<<"  - ObjArray element has address "<< sec->get("Address") <<std::endl;
+    }
 
-	BITPIT_DEBUG_COUT("logger2") << "Debug statement" << std::endl;
-
-	// Log on console and also on file with different verbosities
-	log::manager().create("logger3", true, nProcessors, rank);
-
-	log::cout("logger3").setConsoleVerbosity(log::NORMAL);
-	log::cout("logger3").setFileVerbosity(log::DEBUG);
-
-	log::cout("logger3") << log::priority(log::DEBUG);
-	log::cout("logger3") << log::context("logger3-A");
-	log::cout("logger3") << "1-1\n2--2\n" << std::endl << "3---3" << std::endl;
-	log::cout("logger3") << "4----4\n5-----5\n" << std::endl << "6--";
-	log::cout("logger3") << "----6\n7-------7";
-	log::cout("logger3") << std::endl;
-	log::cout("logger3") << log::priority(log::NORMAL);
-	log::cout("logger3") << log::context("logger3-B");
-	log::cout("logger3") << "1-1\n2--2\n" << std::endl << "3---3" << std::endl;
-	log::cout("logger3") << "4----4\n5-----5\n" << std::endl << "6--";
-	log::cout("logger3") << "----6\n7-------7";
-	log::cout("logger3") << std::endl;
-
-	BITPIT_DEBUG_COUT("logger3") << "Debug statement" << std::endl;
-
-	// Log using logger functions
-	log::manager().create("logger4", true, nProcessors, rank);
-
-	log::cout("logger4").setConsoleVerbosity(log::DEBUG);
-	log::cout("logger4").setFileVerbosity(log::DEBUG);
-
-	log::cout("logger4").setPriority(log::NORMAL);
-	log::cout("logger4").setContext("logger4-B");
-	log::cout("logger4").println("1-1\n2--2\n\n3---3");
-	log::cout("logger4").print("4----4\n5-----5\n\n6--");
-	log::cout("logger4").print("----6\n7-------7");
-	log::cout("logger4").println("");
-	log::cout("logger4").setPriority(log::NORMAL);
-	log::cout("logger4").setContext("logger4-B");
-	log::cout("logger4").println("1-1\n2--2\n\n3---3");
-	log::cout("logger4").print("4----4\n5-----5\n\n6--");
-	log::cout("logger4").print("----6\n7-------7");
-	log::cout("logger4").println("");
-
-	BITPIT_DEBUG_COUT("logger4") << "Debug statement" << std::endl;
-
-	// Test indention
-	log::cout().setConsoleVerbosity(log::DEBUG);
-	log::cout().setFileVerbosity(log::DEBUG);
-
-	log::cout().setPriority(log::NORMAL);
-	log::cout().setContext("indent");
-
-	log::cout().println("012345678912345678901234567890123456789");
-	log::cout().println("<------------------------->");
-	log::cout().setIndentation(4);
-	log::cout().println("<------------------------->");
-	log::cout().setIndentation(4);
-	log::cout().println("<------------------------->");
-	log::cout().setIndentation(-2);
-	log::cout().println("<------------------------->");
-	log::cout().setIndentation(-2);
-	log::cout().println("<------------------------->");
-	log::cout().setIndentation(-4);
-	log::cout().println("<------------------------->");
-
-	log::cout() << "012345678912345678901234567890123456789" << "\n";
-	log::cout() << "<------------------------->" << "\n";
-	log::cout() << log::indent(4);
-	log::cout() << "<------------------------->" << "\n";
-	log::cout() << log::indent(4);
-	log::cout() << "<------------------------->" << "\n";
-	log::cout() << log::indent(-2);
-	log::cout() << "<------------------------->" << "\n";
-	log::cout() << log::indent(-2);
-	log::cout() << "<------------------------->" << "\n";
-	log::cout() << log::indent(-4);
-	log::cout() << "<------------------------->" << "\n";
+    // Write the configuration file in XML and json
+    std::cout << std::endl;
+    std::cout << "Write configuration file..." << std::endl;
+    root.write("config_test00007_updated.xml");
+    root.write("config_test00007_updated.json");
 
     return 0;
 }
@@ -203,30 +106,34 @@ int subtest_001()
 int main(int argc, char *argv[])
 {
 #if BITPIT_ENABLE_MPI==1
-	MPI_Init(&argc,&argv);
+    MPI_Init(&argc,&argv);
 #else
-	BITPIT_UNUSED(argc);
-	BITPIT_UNUSED(argv);
+    BITPIT_UNUSED(argc);
+    BITPIT_UNUSED(argv);
 #endif
 
-	// Initialize the logger
-	log::manager().initialize(log::COMBINED);
+    // Initialize the logger
+    log::manager().initialize(log::COMBINED);
 
-	// Run the subtests
-	log::cout() << "Testing logger" << std::endl;
+#if HAS_RAPIDJSON_LIB
+    // Run the subtests
+    log::cout() << "Testing json configuration parser" << std::endl;
 
-	int status;
-	try {
-		status = subtest_001();
-		if (status != 0) {
-			return status;
-		}
-	} catch (const std::exception &exception) {
-		log::cout() << exception.what();
-		exit(1);
-	}
+    int status;
+    try {
+        status = subtest_001();
+        if (status != 0) {
+            return status;
+        }
+    } catch (const std::exception &exception) {
+        log::cout() << exception.what()<<std::endl;
+        exit(1);
+    }
+#else
+    log::cout() << "Skipping test: configuration parser was compiled without JSON support." << std::endl;
+#endif
 
 #if BITPIT_ENABLE_MPI==1
-	MPI_Finalize();
+    MPI_Finalize();
 #endif
 }
